FEAT: First version of split directories
