+++ CMakeLists.txt
@@ -11,9 +11,15 @@ project(ERDCPP)
 set(LIBSODIUM_INCLUDE_PATH "/usr/local/include")
 set(LIBSODIUM_LIB_PATH "/usr/local/lib")
 
+include_directories(${PROJECT_SOURCE_DIR}/include)
+include_directories(${PROJECT_SOURCE_DIR}/external)
+#include_directories(include/*)
+#include_directories(external/*)
 include_directories (${LIBSODIUM_INCLUDE_PATH})
 find_library(LIBSODIUM_LIBRARY libsodium.so.23.3.0 ${LIBSODIUM_LIB_PATH} REQUIRED)
 
 
 add_subdirectory(src)
 add_subdirectory(tests)
+add_subdirectory(cli)
+add_subdirectory(external)
