+++ external/CMakeLists.txt
@@ -1,0 +1,10 @@
+file(GLOB external
+	json/json.hpp
+	bech32/bech32.h bech32/bech32.cpp
+)
+
+add_library(external ${external})
+	
+	
+
+
