+++ external/bech32/bech32.cpp
