+++ include/account/address.h
@@ -4,7 +4,7 @@
 #include <string>
 #include <vector>
 
-#include "internal.h"
+#include "../internal/internal.h"
 
 class Address
 {
