--- src/InputHandler/ext.h
@@ -1,12 +1,0 @@
-#ifndef EXTINPUTHANDLER_H
-#define EXTINPUTHANDLER_H
-
-#include "signer.h"
-#include "transaction.h"
-#include "ArgHandler/arghandler.h"
-#include "PemHandler/pemhandler.h"
-#include "JsonHandler/jsonhandler.h"
-#include "InputWrappers/pemhandler_input_wrapper.h"
-#include "InputWrappers/jsonhandler_input_wrapper.h"
-
-#endif
