+++ src/account/account.cpp
@@ -1,4 +1,4 @@
-#include "account.h"
+#include "../../include/account/account.h"
 
 Account::Account(Address const&  address):
   m_address(address),
