+++ src/utils/CMakeLists.txt
@@ -1,10 +1,7 @@
 file(GLOB utils
  base64.h base64.cpp
- bech32.h bech32.cpp
- bites.h bites.cpp
- fifo_map.hpp
+ bits.h bits.cpp
  hex.h hex.cpp
- json.hpp
  params.h
  strchr.h
 )
