+++ src/utils/bits.cpp
@@ -1,4 +1,4 @@
-#include "bites.h"
+#include "bits.h"
 
 namespace util
 {
