+++ src/utils/ext.h
@@ -3,11 +3,8 @@
 
 #include "hex.h"
 #include "base64.h"
-#include "bech32.h"
-#include "bites.h"
+#include "bits.h"
 #include "strchr.h"
-#include "json.hpp"
-#include "fifo_map.hpp"
 #include "params.h"
 
 #endif
