+++ src/utils/hex.cpp
