+++ tests/test_utils/CMakeLists.txt
@@ -1,3 +1,5 @@
+include_directories (${PROJECT_SOURCE_DIR}/src/utils)
+
 add_executable(test_utils test_utils.cpp)
 
 target_link_libraries (test_utils PUBLIC gtest_main)
